@@ -3,7 +3,7 @@
     "type": "library",
     "description": "Caching library offering an object-oriented API for many cache backends",
     "keywords": ["cache", "caching"],
-    "homepage": "http://www.doctrine-project.org",
+    "homepage": "https://www.doctrine-project.org",
     "license": "MIT",
     "authors": [
         {"name": "Guilherme Blanco", "email": "guilhermeblanco@gmail.com"},
@@ -13,12 +13,17 @@
         {"name": "Johannes Schmitt", "email": "schmittjoh@gmail.com"}
     ],
     "require": {
-        "php": "~5.5|~7.0"
+        "php": "~7.1"
     },
     "require-dev": {
-        "phpunit/phpunit":         "~4.8|~5.0",
-        "satooshi/php-coveralls":  "~0.6",
-        "predis/predis":           "~1.0"
+        "alcaeus/mongo-php-adapter": "^1.1",
+        "mongodb/mongodb": "^1.1",
+        "phpunit/phpunit": "^7.0",
+        "predis/predis":   "~1.0",
+        "doctrine/coding-standard": "^4.0"
+    },
+    "suggest": {
+        "alcaeus/mongo-php-adapter": "Required to use legacy MongoDB driver"
     },
     "conflict": {
         "doctrine/common": ">2.2,<2.4"
@@ -31,7 +36,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.6.x-dev"
+            "dev-master": "1.8.x-dev"
         }
     }
 }
