@@ -1,36 +1,52 @@
 {
     "name": "doctrine/common",
     "type": "library",
-    "description": "Common Library for Doctrine projects",
-    "keywords": ["collections", "spl", "eventmanager", "annotations", "persistence"],
-    "homepage": "http://www.doctrine-project.org",
+    "description": "PHP Doctrine Common project is a library that provides additional functionality that other Doctrine projects depend on such as better reflection support, persistence interfaces, proxies, event system and much more.",
+    "keywords": [
+        "php",
+        "common",
+        "doctrine"
+    ],
+    "homepage": "https://www.doctrine-project.org/projects/common.html",
     "license": "MIT",
     "authors": [
         {"name": "Guilherme Blanco", "email": "guilhermeblanco@gmail.com"},
         {"name": "Roman Borschel", "email": "roman@code-factory.org"},
         {"name": "Benjamin Eberlei", "email": "kontakt@beberlei.de"},
         {"name": "Jonathan Wage", "email": "jonwage@gmail.com"},
-        {"name": "Johannes Schmitt", "email": "schmittjoh@gmail.com"}
+        {"name": "Johannes Schmitt", "email": "schmittjoh@gmail.com"},
+        {"name": "Marco Pivetta", "email": "ocramius@gmail.com"}
     ],
     "require": {
-        "php": "~5.6|~7.0",
-        "doctrine/inflector": "1.*",
-        "doctrine/cache": "1.*",
-        "doctrine/collections": "1.*",
-        "doctrine/lexer": "1.*",
-        "doctrine/annotations": "1.*"
+        "php": "^7.1",
+        "doctrine/inflector": "^1.0",
+        "doctrine/cache": "^1.0",
+        "doctrine/collections": "^1.0",
+        "doctrine/lexer": "^1.0",
+        "doctrine/annotations": "^1.0",
+        "doctrine/event-manager": "^1.0",
+        "doctrine/reflection": "^1.0",
+        "doctrine/persistence": "^1.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.4.6"
+        "phpunit/phpunit": "^6.3",
+        "doctrine/coding-standard": "^1.0",
+        "squizlabs/php_codesniffer": "^3.0",
+        "symfony/phpunit-bridge": "^4.0.5"
     },
     "autoload": {
         "psr-4": {
             "Doctrine\\Common\\": "lib/Doctrine/Common"
         }
     },
+    "autoload-dev": {
+        "psr-4": {
+            "Doctrine\\Tests\\": "tests/Doctrine/Tests"
+        }
+    },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.7.x-dev"
+            "dev-master": "2.10.x-dev"
         }
     }
 }
