@@ -12,30 +12,35 @@
     "authors": [
         {
             "name": "Sebastian Bergmann",
-            "email": "sb@sebastian-bergmann.de",
+            "email": "sebastian@phpunit.de",
             "role": "lead"
         }
     ],
     "support": {
-        "issues": "https://github.com/sebastianbergmann/php-code-coverage/issues",
-        "irc": "irc://irc.freenode.net/phpunit"
+        "issues": "https://github.com/sebastianbergmann/php-code-coverage/issues"
     },
+    "config": {
+        "optimize-autoloader": true,
+        "sort-packages": true
+    },
+    "prefer-stable": true,
     "require": {
-        "php": ">=5.3.3",
-        "phpunit/php-file-iterator": "~1.3",
-        "phpunit/php-token-stream": "~1.3",
-        "phpunit/php-text-template": "~1.2",
-        "sebastian/environment": "^1.3.2",
-        "sebastian/version": "~1.0"
+        "php": "^7.0",
+        "ext-dom": "*",
+        "ext-xmlwriter": "*",
+        "phpunit/php-file-iterator": "^1.4.2",
+        "phpunit/php-token-stream": "^2.0.1",
+        "phpunit/php-text-template": "^1.2.1",
+        "sebastian/code-unit-reverse-lookup": "^1.0.1",
+        "sebastian/environment": "^3.0",
+        "sebastian/version": "^2.0.1",
+        "theseer/tokenizer": "^1.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4",
-        "ext-xdebug": ">=2.1.4"
+        "phpunit/phpunit": "^6.0"
     },
     "suggest": {
-        "ext-dom": "*",
-        "ext-xdebug": ">=2.2.1",
-        "ext-xmlwriter": "*"
+        "ext-xdebug": "^2.5.5"
     },
     "autoload": {
         "classmap": [
@@ -44,7 +49,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.2.x-dev"
+            "dev-master": "5.3.x-dev"
         }
     }
 }
