@@ -11,22 +11,29 @@
     "authors": [
         {
             "name": "Sebastian Bergmann",
-            "email": "sb@sebastian-bergmann.de",
+            "email": "sebastian@phpunit.de",
             "role": "lead"
         }
     ],
     "support": {
-        "issues": "https://github.com/sebastianbergmann/phpunit-mock-objects/issues",
-        "irc": "irc://irc.freenode.net/phpunit"
+        "issues": "https://github.com/sebastianbergmann/phpunit-mock-objects/issues"
     },
+    "prefer-stable": true,
     "require": {
-        "php": ">=5.3.3",
-        "phpunit/php-text-template": "~1.2",
-        "doctrine/instantiator": "^1.0.2",
-        "sebastian/exporter": "~1.2"
+        "php": "^7.0",
+        "phpunit/php-text-template": "^1.2.1",
+        "doctrine/instantiator": "^1.0.5",
+        "sebastian/exporter": "^3.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.4"
+        "phpunit/phpunit": "^6.5.11"
+    },
+    "conflict": {
+        "phpunit/phpunit": "<6.0"
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "sort-packages": true
     },
     "suggest": {
         "ext-soap": "*"
@@ -43,7 +50,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.3.x-dev"
+            "dev-master": "5.0.x-dev"
         }
     }
 }
