@@ -1195,12 +1195,11 @@
  * @section running Running tests
  * You can run both Simpletest and PHPUnit tests by enabling the core Testing
  * module (core/modules/simpletest). Once that module is enabled, tests can be
- * run using the core/scripts/run-tests.sh script, using
- * @link https://www.drupal.org/project/drush Drush @endlink, or from the
- *   Testing module user interface.
+ * run on through the Testing module's user interface or on the command line.
  *
- * PHPUnit tests can also be run from the command line, using the PHPUnit
- * framework. See https://www.drupal.org/node/2116263 for more information.
+ * See @link https://www.drupal.org/node/2116263 Running tests with PHPUnit
+ * binary @endlink or @link https://www.drupal.org/node/645286 Running tests
+ * with run-tests.sh @endlink for more information.
  * @}
  */
 
