@@ -15,7 +15,7 @@ interface DeriverInterface {
    * @param string $derivative_id
    *   The derivative id. The id must uniquely identify the derivative within a
    *   given base plugin, but derivative ids can be reused across base plugins.
-   * @param mixed $base_plugin_definition
+   * @param array|\Drupal\Component\Plugin\Definition\PluginDefinitionInterface $base_plugin_definition
    *   The definition of the base plugin from which the derivative plugin
    *   is derived. It is maybe an entire object or just some array, depending
    *   on the discovery mechanism.
