@@ -12,10 +12,9 @@ interface LogMessageParserInterface {
    *
    * For a value to be considered as a placeholder should be in the following
    * formats:
-   *   - PSR3 format:
-   *     @see https://github.com/php-fig/fig-standards/blob/master/accepted/PSR-3-logger-interface.md#12-message
-   *   - Drupal specific string placeholder format:
-   *     @see \Drupal\Component\Render\FormattableMarkup
+   *   - @link https://github.com/php-fig/fig-standards/blob/master/accepted/PSR-3-logger-interface.md#12-message The PSR3 format @endlink
+   *   - The Drupal specific string placeholder format, described in
+   *     \Drupal\Component\Render\FormattableMarkup
    *
    * Values in PSR3 format will be transformed to
    * \Drupal\Component\Render\FormattableMarkup format.
@@ -29,6 +28,8 @@ interface LogMessageParserInterface {
    *
    * @return array
    *   An array of the extracted message placeholders.
+   *
+   * @see \Drupal\Component\Render\FormattableMarkup
    */
   public function parseMessagePlaceholders(&$message, array &$context);
 
