@@ -2,7 +2,7 @@
 
 namespace Drupal\config_translation\Plugin\migrate\source\d6;
 
-use Drupal\migrate_drupal\Plugin\migrate\source\DrupalSqlBase;
+use Drupal\user\Plugin\migrate\source\ProfileField;
 
 /**
  * Gets i18n strings profile field source from database.
@@ -12,16 +12,16 @@ use Drupal\migrate_drupal\Plugin\migrate\source\DrupalSqlBase;
  *   source_module = "i18nprofile"
  * )
  */
-class ProfileFieldTranslation extends DrupalSqlBase {
+class ProfileFieldTranslation extends ProfileField {
 
   /**
    * {@inheritdoc}
    */
   public function query() {
-    $query = $this->select('profile_fields', 'pf')
-      ->fields('pf', ['fid', 'name'])
-      ->fields('i18n', ['property'])
-      ->fields('lt', ['lid', 'translation', 'language']);
+    $query = parent::query();
+    $query->fields('i18n', ['property'])
+      ->fields('lt', ['lid', 'translation', 'language'])
+      ->isNotNull('language');
     $query->leftJoin('i18n_strings', 'i18n', 'i18n.objectid = pf.name');
     $query->leftJoin('locales_target', 'lt', 'lt.lid = i18n.lid');
     return $query;
@@ -31,23 +31,22 @@ class ProfileFieldTranslation extends DrupalSqlBase {
    * {@inheritdoc}
    */
   public function fields() {
-    return [
-      'fid' => $this->t('Profile field ID.'),
+    $fields = [
       'lid' => $this->t('Locales target language ID.'),
       'language' => $this->t('Language for this field.'),
       'translation' => $this->t('Translation of either the title or explanation.'),
     ];
+    return parent::fields() + $fields;
   }
 
   /**
    * {@inheritdoc}
    */
   public function getIds() {
-    $ids['fid']['type'] = 'integer';
     $ids['language']['type'] = 'string';
     $ids['lid']['type'] = 'integer';
     $ids['lid']['alias'] = 'lt';
-    return $ids;
+    return parent::getIds() + $ids;
   }
 
 }
