@@ -12,7 +12,7 @@ use Drupal\simpletest\WebTestBase;
  * database tables. Therefore, we ensure that tests which are not set up
  * properly are skipped.
  *
- * @group simpletest
+ * @group WebTestBase
  * @see \Drupal\simpletest\WebTestBase
  */
 class BrokenSetUpTest extends WebTestBase {
