@@ -106,7 +106,7 @@ class Workspace extends ContentEntityBase implements WorkspaceInterface {
 
     $fields['created'] = BaseFieldDefinition::create('created')
       ->setLabel(new TranslatableMarkup('Created'))
-      ->setDescription(new TranslatableMarkup('The time that the workspaces was created.'));
+      ->setDescription(new TranslatableMarkup('The time that the workspace was created.'));
 
     return $fields;
   }
