@@ -44,14 +44,11 @@ class EnvironmentTest extends TestCase {
    *   required and memory_limit, and the expected return value.
    */
   public function providerTestCheckMemoryLimit() {
-    $memory_limit = ini_get('memory_limit');
-    $twice_avail_memory = ($memory_limit * 2) . 'MB';
-
     return [
       // Minimal amount of memory should be available.
       ['30MB', NULL, TRUE],
-      // Exceed a custom (unlimited) memory limit.
-      [$twice_avail_memory, -1, TRUE],
+      // Test an unlimited memory limit.
+      ['9999999999YB', -1, TRUE],
       // Exceed a custom memory limit.
       ['30MB', '16MB', FALSE],
       // Available = required.
